step 5, key fixed!

+++ database/migrations/2021_03_20_164611_create_blog_posts_table.php
@@ -16,8 +16,8 @@ class CreateBlogPostsTable extends Migration
         Schema::create('blog_posts', function (Blueprint $table) {
             $table->increments('id');
 
-            $table->integer('category_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->unsignedInteger('category_id')->unsigned();
+            $table->unsignedBigInteger('user_id')->unsigned();
 
             $table->string('slug')->unique();
             $table->string('title');
